% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_bezier.R
\name{analyze_bezier}
\alias{analyze_bezier}
\title{Analyze bezier curves of any order}
\usage{
analyze_bezier(cp_x, cp_y)
}
\arguments{
\item{cp_x}{the x-axis coordinate of the control points}

\item{cp_y}{the y-axis coordinate of the control points}
}
\value{

}
\description{
Analyze bezier curves of any order
}
