% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_svg.R
\name{analyze_svg}
\alias{analyze_svg}
\title{This will take any svg file and show the control point skeleton.}
\usage{
analyze_svg(svg_path)
}
\arguments{
\item{svg_path}{svg file path}
}
\value{

}
\description{
This will take any svg file and show the control point skeleton.
}
