% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_bezier.R
\name{animate_bezier}
\alias{animate_bezier}
\title{Animate bezier curves of any order}
\usage{
animate_bezier(cp_x, cp_y)
}
\arguments{
\item{cp_x}{the x-axis coordinate of the control points}

\item{cp_y}{the y-axis coordinate of the control points}
}
\value{

}
\description{
Animate bezier curves of any order
}
