% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_svg.R
\name{animate_svg}
\alias{animate_svg}
\title{This function analyzes and animate svg files.}
\usage{
animate_svg(svg_path, des_path)
}
\arguments{
\item{svg_path}{the path for the svg file}
}
\value{

}
\description{
This function analyzes and animate svg files.
}
