\name{bezier-package}
\alias{bezier-package}
\docType{package}
\title{Bezier Curve and Spline Toolkit}
\description{The bezier package is a toolkit for working with Bezier curves and splines. The package provides functions for point generation, arc length estimation, degree elevation and curve fitting.}
\details{
\tabular{ll}{
Package: \tab bezier\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2014-03-05\cr
License: \tab GPL-2\cr
}
}
\author{Aaron Olsen}
\keyword{ package }