% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_base.R
\name{draw_base}
\alias{draw_base}
\title{Draw a bezier curve given any number of control points}
\usage{
draw_base(x_p, y_p, vp)
}
\arguments{
\item{x_p}{control points for the x-axis}

\item{y_p}{control points for the y-axis}

\item{vp}{viewport passed to plot the drawing}
}
\description{
Draw a bezier curve given any number of control points
}
