% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_terms.R
\name{make_terms}
\alias{make_terms}
\title{This function creates the basis equation for the bezier curves for any order}
\usage{
make_terms(n, cp)
}
\arguments{
\item{n}{the order of the basis equation}

\item{cp}{the coefficients}
}
\value{

}
\description{
This function creates the basis equation for the bezier curves for any order
}
