% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_svg.R
\name{parsed_svg}
\alias{parsed_svg}
\title{Parseing svg file}
\usage{
parsed_svg(path)
}
\arguments{
\item{path}{the path for the SVG file.}
}
\value{
dataframe of the parsed SVG file
}
\description{
Parseing svg file
}
