% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_base.R
\name{pre_drawing}
\alias{pre_drawing}
\title{Set up the drawing area for the bezier curve}
\usage{
pre_drawing(x_min, x_max, y_min, y_max)
}
\arguments{
\item{x_min}{the minimum value for all points on x-axis}

\item{x_max}{the maximum value for all points on x-axis}

\item{y_min}{the minimum value for all points on y-axis}

\item{y_max}{the maximum value for all points on y-axis}
}
\value{

}
\description{
Set up the drawing area for the bezier curve
}
